analytics/requirements.txt
pymongo
pandas
numpy
matplotlib
python-dotenv
